scripts: move most scripts related to demo generation to the scripts directory
