+++ package.json
@@ -73,8 +73,8 @@
     "check:appveyor": "npm run check:types && npm run lint && npm run lint:demo && npm run lint:tests && npm run test:appveyor",
     "check:travis": "npm run check:types && npm run lint && npm run lint:demo && npm run lint:tests && npm run test:travis",
     "check:types": "tsc --noEmit --project .",
-    "demo": "cross-env RXP_METAPLAYLIST=true webpack --progress --config webpack-demo.config.js",
-    "demo:min": "cross-env RXP_MINIFY=true RXP_METAPLAYLIST=true webpack --progress --config webpack-demo.config.js",
+    "demo": "cross-env RXP_METAPLAYLIST=true node ./scripts/generate_full_demo.js",
+    "demo:min": "cross-env RXP_METAPLAYLIST=true RXP_MINIFY=true webpack --progress --config webpack-demo.config.js",
     "demo:watch": "webpack --progress --config webpack-demo.config.js -w",
     "doc": "rm -rf doc/generated; node ./scripts/doc-generator/index.js doc/ doc/generated",
     "info": "npm-scripts-info",
@@ -83,8 +83,8 @@
     "lint:tests": "eslint tests/**/*.js",
     "prepare:modular": "tsc -p tsconfig.modules.json",
     "prepublishOnly": "npm run build:modular",
-    "standalone": "cross-env RXP_ENV=development node ./demo/standalone/server.js",
-    "start": "cross-env RXP_ENV=development RXP_METAPLAYLIST=true node ./demo/full/server.js",
+    "standalone": "cross-env RXP_METAPLAYLIST=true RXP_ENV=development node ./scripts/run_standalone_demo.js",
+    "start": "cross-env RXP_METAPLAYLIST=true RXP_ENV=development node ./scripts/run_full_demo.js",
     "test": "npm run test:unit && npm run test:integration",
     "test:appveyor": "npm run test:unit && npm run test:integration:appveyor && npm run test:memory",
     "test:integration": "node tests/integration/run.js --bchromehl --bfirefoxhl",
