+++ scripts/generate_documentation_list.js
@@ -1,3 +1,4 @@
+#!/usr/bin/env node
 /* eslint-env node */
 
 /**
